Refactor data handling in DedalusDataset2D class by removing unnecessary float conversion and ensuring consistent data type after permutation.

+++ utils/griddataset.py
@@ -1142,7 +1142,7 @@ class DedalusDataset2D(Dataset):
                 timestep_aug = np.tile(timestep_data, (H, W, 1)).transpose(2, 0, 1)  # (T, H, W)
                 
                 # Stack all data
-                data = np.stack([vorticity_data, streamfunction_data], axis=1).float()  # (T, C, H, W)
+                data = np.stack([vorticity_data, streamfunction_data], axis=1) # (T, C, H, W)
                 
             except (OSError, KeyError, IndexError):
                 # Fallback: use the main virtual dataset file
@@ -1188,7 +1188,7 @@ class DedalusDataset2D(Dataset):
             data = self.downsample_x(data, H//self.downsample[0])
         
         # Reshape to (H, W, T, C)
-        data = data.permute(2, 3, 0, 1) # (T, C, H, W) -> (H, W, T, C)
+        data = data.permute(2, 3, 0, 1).astype(torch.float32) # (T, C, H, W) -> (H, W, T, C)
         x = data[..., :self.t_in, :]
         y = data[..., self.t_in:self.t_in + self.t_out, :]
         return x, y
